refactor(ui): Consolidate properties panel components
Merge separate property panel component files into a single file to improve
maintainability and reduce code duplication. Components include:

- Collapsible panels
- Dropdown menus
- Hover cards
- Tabs
- Other shared UI primitives

No functional changes or breaking updates.
